+++ board/iotdk/drivers/ip/subsystem/iic/dfss_iic_obj.h
@@ -1,5 +1,5 @@
 /* ------------------------------------------
- * Copyright (c) 2017, Synopsys, Inc. All rights reserved.
+ * Copyright (c) 2018, Synopsys, Inc. All rights reserved.
 
  * Redistribution and use in source and binary forms, with or without modification,
  * are permitted provided that the following conditions are met:
@@ -27,25 +27,48 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
 --------------------------------------------- */
-#ifndef _DW_I2S_OBJ_H_
-#define _DW_I2S_OBJ_H_
 
-#include "ip/ip_hal/inc/dev_i2s.h"
+#ifndef _DFSS_IIC_OBJ_H_
+#define _DFSS_IIC_OBJ_H_
 
-#define DW_I2S_0_ID		0	/*!< I2S 0 ID macro */
-#define DW_I2S_1_ID		1	/*!< I2S 1 ID macro */
+#include "io_config.h"
+#include "ip/ip_hal/inc/dev_iic.h"
 
-#define USE_DW_I2S_0		1	/*!< enable use DesignWare I2S 0 as TX */
-#define USE_DW_I2S_1		1	/*!< enable use DesignWare I2S 1 as TX */
+
+#define DFSS_IIC_NUM		(3)	/*!< DFSS IIC valid number */
+
+
+#define DFSS_IIC_0_ID		0	/*!< IIC 0 ID macro */
+#define DFSS_IIC_1_ID		1	/*!< IIC 1 ID macro */
+
+
+#ifdef IO_I2C_MST0_PRESENT
+#define USE_DFSS_IIC_0		1	/*!< enable use DFSS IIC 0 */
+#else
+#define USE_DFSS_IIC_0		0	/*!< enable use DFSS IIC 0 */
+#endif
+
+#ifdef IO_I2C_MST1_PRESENT
+#define USE_DFSS_IIC_1		1	/*!< enable use DFSS IIC 1 */
+#else
+#define USE_DFSS_IIC_1		0	/*!< enable use DFSS IIC 1 */
+#endif
+
+#ifdef IO_I2C_MST2_PRESENT
+#define USE_DFSS_IIC_2		1	/*!< enable use DFSS IIC 2 */
+#else
+#define USE_DFSS_IIC_2		0	/*!< enable use DFSS IIC 2 */
+#endif
+
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-extern DEV_I2S_PTR i2s_get_dev(int32_t i2s_id);
+extern void dfss_iic_all_install(void);
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif /* _DW_I2S_OBJ_H_ */
+#endif /* _DFSS_IIC_OBJ_H_ */
