+++ device/ip/ip_hal/inc/ioctl.h
@@ -1,0 +1,63 @@
+/* ==========================================================================
+* Synopsys DesignWare Sensor and Control IP Subsystem IO Software Driver and
+* documentation (hereinafter, "Software") is an Unsupported proprietary work
+* of Synopsys, Inc. unless otherwise expressly agreed to in writing between
+* Synopsys and you.
+*
+* The Software IS NOT an item of Licensed Software or Licensed Product under
+* any End User Software License Agreement or Agreement for Licensed Product
+* with Synopsys or any supplement thereto. You are permitted to use and
+* redistribute this Software in source and binary forms, with or without
+* modification, provided that redistributions of source code must retain this
+* notice. You may not view, use, disclose, copy or distribute this file or
+* any information contained herein except pursuant to this license grant from
+* Synopsys. If you do not agree with this notice, including the disclaimer
+* below, then you are not authorized to use the Software.
+*
+* THIS SOFTWARE IS BEING DISTRIBUTED BY SYNOPSYS SOLELY ON AN "AS IS" BASIS
+* AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+* IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+* ARE HEREBY DISCLAIMED. IN NO EVENT SHALL SYNOPSYS BE LIABLE FOR ANY DIRECT,
+* INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+* (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+* SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+* CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+* LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
+* OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH
+* DAMAGE.
+* ========================================================================== */
+/*==========================================================================
+* Library DW_DFSS-1.1.6
+* ========================================================================== */
+
+#ifndef IOCTL_H_
+#define IOCTL_H_
+
+
+typedef void (*IO_CB_FUNC) (uint32_t);
+
+typedef struct {
+    IO_CB_FUNC cb;
+} io_cb_t;
+
+// IO devices common control commands
+#define     IO_SET_CB_RX			(0x01)
+#define     IO_SET_CB_TX 			(0x02)
+#define     IO_SET_CB_ERR			(0x03)
+
+
+/* the following are necessary definitions in embARC_OSP */
+#include "arc_exception.h"
+
+#undef _Interrupt
+#define _Interrupt
+
+#undef _Uncached
+#define _Uncached volatile
+
+#define _setvecti(x, y)  {int_handler_install(x, y); int_enable(x);}
+
+#undef __Xdmac
+
+
+#endif
